refactor(client): 更新账户信息 API 版本并优化相关功能
- 将 Get Account API 版本从 v1 更新到 v3
- 修改 GoodTillDate 字段类型从 int 改为 int64
- 新增 websocket 账户信息查询功能

+++ trade.go
@@ -889,7 +889,7 @@ type GetOrderResponse struct {
 	PriceProtect  bool         `json:"priceProtect"`
 	PriceMatch    PriceMatch   `json:"priceMatch"`
 	StpMode       STPMode      `json:"selfTradePreventionMode"`
-	GoodTillDate  int          `json:"goodTillDate"`
+	GoodTillDate  int64        `json:"goodTillDate"`
 }
 
 // Get All Order 查询所有订单 (GET /fapi/v1/allOrders)
@@ -989,7 +989,7 @@ type NewAllOrdersResponse struct {
 	PriceProtect  bool         `json:"priceProtect"`
 	PriceMatch    PriceMatch   `json:"priceMatch"`
 	StpMode       STPMode      `json:"selfTradePreventionMode"`
-	GoodTillDate  int          `json:"goodTillDate"`
+	GoodTillDate  int64        `json:"goodTillDate"`
 }
 
 // Get open Order 查询当前挂单 (GET /fapi/v1/openOrder)
